+++ instrumentation/hibernate/hibernate-reactive-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/hibernate/reactive/v1_0/stage/StageSessionImplInstrumentation.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.hibernate.reactive.v1_0;
+package io.opentelemetry.javaagent.instrumentation.hibernate.reactive.v1_0.stage;
 
 import static net.bytebuddy.matcher.ElementMatchers.named;
 import static net.bytebuddy.matcher.ElementMatchers.namedOneOf;
@@ -18,45 +18,30 @@ import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 
-public class StageSessionFactoryInstrumentation implements TypeInstrumentation {
+public class StageSessionImplInstrumentation implements TypeInstrumentation {
   @Override
   public ElementMatcher<TypeDescription> typeMatcher() {
-    return named("org.hibernate.reactive.stage.impl.StageSessionFactoryImpl");
+    return namedOneOf(
+        "org.hibernate.reactive.stage.impl.StageSessionImpl",
+        "org.hibernate.reactive.stage.impl.StageStatelessSessionImpl");
   }
 
   @Override
   public void transform(TypeTransformer transformer) {
     transformer.applyAdviceToMethod(
-        namedOneOf("withSession", "withStatelessSession").and(takesArgument(0, Function.class)),
-        this.getClass().getName() + "$Function0Advice");
-    transformer.applyAdviceToMethod(
-        namedOneOf("withSession", "withStatelessSession").and(takesArgument(1, Function.class)),
-        this.getClass().getName() + "$Function1Advice");
-    transformer.applyAdviceToMethod(
-        namedOneOf("openSession", "openStatelessSession").and(returns(CompletionStage.class)),
-        this.getClass().getName() + "$OpenSessionAdvice");
+        named("withTransaction")
+            .and(takesArgument(0, Function.class).and(returns(CompletionStage.class))),
+        this.getClass().getName() + "$WithTransactionAdvice");
   }
 
   @SuppressWarnings("unused")
-  public static class Function0Advice {
+  public static class WithTransactionAdvice {
     @Advice.OnMethodEnter(suppress = Throwable.class)
     public static void onEnter(
         @Advice.Argument(value = 0, readOnly = false) Function<?, ?> function) {
       function = FunctionWrapper.wrap(function);
     }
-  }
 
-  @SuppressWarnings("unused")
-  public static class Function1Advice {
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void onEnter(
-        @Advice.Argument(value = 1, readOnly = false) Function<?, ?> function) {
-      function = FunctionWrapper.wrap(function);
-    }
-  }
-
-  @SuppressWarnings("unused")
-  public static class OpenSessionAdvice {
     @Advice.OnMethodExit(suppress = Throwable.class)
     public static void onExit(@Advice.Return(readOnly = false) CompletionStage<?> completionStage) {
       completionStage = CompletionStageWrapper.wrap(completionStage);
